moved arrow/parquet parsers to it's own file and reverted changes
